requirements.txt
requests==2.31.0
sendgrid==6.6.0
shapely==2.0.4
python-dotenv==1.0.1
PyYAML==6.0.1
termcolor==2.4.0
boto3==1.34.98
pytz==2024.1